fix(db): fix gorm v2 semantic regressions and batch-insert optimizations
The mechanical v1→v2 migration (178272476) converted imports, tags,
error patterns, and hook signatures, but did not address gorm v2's
shared-Statement semantics, stricter WHERE guards, or dialect-name
changes. This commit fixes all discovered runtime regressions and
applies CreateInBatches optimizations.

Session isolation (gorm v2 shared Statement):
- sfdb/rule_op.go: addGroupsForRule after db.Create uses NewDB session
  to avoid Table pollution; CreateOrUpdateRuleWithGroup same fix
- yakit/syntaxflow_rule.go: FilterSyntaxFlowRule returns Session for
  clone>=2 so callers' repeated finishers (Count/Find/Pluck) don't
  accumulate; UpdateSyntaxFlowRule adds WHERE rule_name=? (v2 requires
  explicit WHERE for Updates on empty primary key)
- yakit/syntaxflow_result.go: FilterSyntaxFlowResult returns Session
- bizhelper/import.go: NewTableImportTool uses Session to isolate
  createOrUpdateByUniqueIndex loop (fixes ambiguous column: rule_id)
- yakit/ssa_risk_disposals.go: UpdateSSARiskDisposals Save per-disposal
  uses Session (fixes WHERE conditions required from Model overwriting
  primary key)
- yakit/knowledge_base.go: DeleteKnowledgeBase uses per-table Session
  for cross-table Delete (fixes no such column: knowledge_base_id);
  UpdateKnowledgeBaseEntryByHiddenIndex isolates Count vs Updates
  (fixes ambiguous column name: hidden_index)
- yakit/ai_event_stream_buffer.go: First calls use Session to avoid
  Statement pollution from upstream Where; fix dialect name check
  from v1 "sqlite3" to v2 "sqlite"
- yakit/ai_yak_tool.go: all First/Save/Create/Updates/FirstOrCreate
  use Session/NewDB isolation (fixes ambiguous column: name and
  WHERE conditions required)
- yakit/httpflow_tree_helper.go: Order(clause.OrderBy{}) generates
  empty "ORDER BY " causing incomplete input; replaced with plain
  Limit(-1).Offset(-1)

Schema tag fix:
- schema/vector_store.go: CollectionUUID had uniqueIndex which v1
  silently ignored but v2 enforces, causing UNIQUE constraint failed
  on multi-doc collections; changed to index

Test fixture fixes:
- option_plugin_test.go: BeforeSave(nil) to match v2 hook signature
- 9 test files: int64 Count assertions corrected (Count returns int64,
  not int)

Batch-insert optimizations (gorm v2 CreateInBatches):
- ssa/database_cache_source.go: per-record Create → CreateInBatches(500)
- ssa/ssadb/index.go: per-record Create → CreateInBatches(500)
- grpc_httpflow.go: websocket frame per-record Create →
  CreateInBatches(500)

Verified: go build ./... passes; schema/consts/ssadb/bizhelper/sfdb/
yakit/vectorstore/sfweb/grpc MUSTPASS (SSA/SyntaxFlow/Language)/
coreplugin/static_analyzer tests all pass.

+++ common/ai/rag/plugins_rag/test/plugins_rag_test.go
@@ -95,7 +95,7 @@ func createTestPlugins(t *testing.T) {
 	// 验证插件已创建
 	var count int64
 	db.Model(&schema.YakScript{}).Where("script_name LIKE ?", "test_rag_plugin%").Count(&count)
-	assert.Equal(t, 3, count, "应该创建3个测试插件")
+	assert.Equal(t, int64(3), count, "应该创建3个测试插件")
 }
 
 // 清理测试用插件
