+++ common/yakgrpc/grpc_syntaxflow_result_test.go
@@ -351,7 +351,7 @@ func TestGRPCMUSTPASS_SyntaxFlow_ResultDelete(t *testing.T) {
 		riskCount := int64(0)
 		err = consts.GetGormProjectDatabase().Model(&schema.Risk{}).Where("result_id = ?", resultID1).Count(&riskCount).Error
 		require.NoError(t, err)
-		require.Equal(t, 0, riskCount)
+		require.Equal(t, int64(0), riskCount)
 	})
 
 	t.Run("test delete contain risk", func(t *testing.T) {
